fix(admin): Correção da estrura das pastas
