+++ src/constants/rotas.ts
@@ -4,23 +4,23 @@
 export const ROTAS = {
   CADASTRO: "/cadastro",
   LOGIN: "/login",
-  ESTATISTICAS: "/admin/admin/estatisticas",
-  USUARIOS: "/admin/admin/usuarios",
-  ADICIONAR_USUARIOS: "/admin/admin/usuarios/adicionar",
-  EDITAR_USUARIOS: "/admin/admin/usuarios/editar",
-  CATEGORIAS: "/admin/blog/categorias",
-  POSTS: "/admin/blog/posts",
-  ADICIONAR_POSTS: "/admin/blog/posts/adicionar",
-  EDITAR_POSTS: "/admin/blog/posts/editar",
-  CAPITULOS: "/admin/principal/capitulos/",
-  ADICIONAR_CAPITULO: "/admin/principal/capitulos/adicionar",
-  EDITAR_CAPITULO: "/admin/principal/capitulos/editar",
-  DASHBOARD: "/admin/principal/dashboard",
-  GENEROS: "/admin/principal/generos",
-  OBRAS: "/admin/principal/obras",
-  ADICIONAR_OBRA: "/admin/principal/obras/adicionar",
-  EDITAR_OBRA: "/admin/principal/obras/editar",
-  VOLUMES: "/admin/principal/volumes",
-  ADICIONAR_VOLUME: "/admin/principal/volumes/adicionar",
-  EDITAR_VOLUME: "/admin/principal/volumes/editar",
+  ESTATISTICAS: "/admin/estatisticas",
+  USUARIOS: "/admin/usuarios",
+  ADICIONAR_USUARIOS: "/admin/usuarios/adicionar",
+  EDITAR_USUARIOS: "/admin/usuarios/editar",
+  CATEGORIAS: "/admin/categorias",
+  POSTS: "/admin/posts",
+  ADICIONAR_POSTS: "/admin/posts/adicionar",
+  EDITAR_POSTS: "/admin/posts/editar",
+  CAPITULOS: "/admin/capitulos/",
+  ADICIONAR_CAPITULO: "/admin/capitulos/adicionar",
+  EDITAR_CAPITULO: "/admin/capitulos/editar",
+  DASHBOARD: "/admin",
+  GENEROS: "/admin/generos",
+  OBRAS: "/admin/obras",
+  ADICIONAR_OBRA: "/admin/obras/adicionar",
+  EDITAR_OBRA: "/admin/obras/editar",
+  VOLUMES: "/admin/volumes",
+  ADICIONAR_VOLUME: "/admin/volumes/adicionar",
+  EDITAR_VOLUME: "/admin/volumes/editar",
 };
