--- src/pages/admin/blog/categorias/styles.ts
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-export const Titulo = styled.h1`  
-      color: ${({ theme }) => theme.colors.primaria[500]};
-`;
