--- src/pages/admin/blog/posts/editar/index.tsx
@@ -1,26 +1,0 @@
-import { Navbar } from "@/components/Admin/Heading/Navbar";
-import LayoutAdminMain from "@/components/Layouts/Admin/Main";
-import { ROTAS } from "@/constants/rotas";
-import { BiArrowBack } from "react-icons/bi";
-import { Container, Titulo } from "../styles";
-
-const Editar = () => {
-  return (
-    <LayoutAdminMain>
-      <Navbar
-        title="Editar Post"
-        placeholderSearchbox=""
-        showSearhBox={false}
-        titleLinkNavigation="Voltar"
-        showLinkNavigation={true}
-        urlLinkNavigation={ROTAS.POSTS}
-        icon={<BiArrowBack />}
-      />
-      <Container>
-        <Titulo>Editar Post</Titulo>
-      </Container>
-    </LayoutAdminMain>
-  );
-};
-
-export default Editar;
